refactor: replace int with size_t/uint types for lengths and counts
Systematically replace signed int parameters and fields with
appropriate unsigned types throughout the FIDO2 codebase, eliminating
signed/unsigned mismatch compiler warnings. Where the underlying
expression type cannot be changed directly (e.g. arithmetic on narrow
fields, bit-shift results), explicit casts are added instead.

- Use size_t for buffer lengths, data sizes, and byte counts
- Use uint8_t for protocol IDs, sub-commands, and other narrow fields
- Use uint16_t for U2F return codes (u2f_register, u2f_authenticate,
  u2f_version)
- Use uint32_t for AES buffer lengths and APDU lc parsing

Also includes several related fixes uncovered during the cleanup:
- Fix NEW_PIN_ENC_MAX_SIZE (256 → 80) to match the FIDO2 specification
- Fix double-negation bug in ctap_client_pin_verify_permissions_rp_id
  check in ctap_make_credential, where the boolean result was
  incorrectly compared against CTAP2_OK
- Move hashbuf in ctap_sign_data from caller-supplied to local stack
  variable, removing an aliased buffer argument
- Drop unused credtype parameter from cose_key_add
- Drop unused parameters (CborEncoder* from parse_make_credential,
  appid from u2f_load_key)
- Make u2f_request_ex, u2f_new_keypair, get_user_present static or
  non-static as appropriate to reflect intended visibility
- Add bounds checks before narrowing CBOR integers to uint8_t
- Add assert in decrypt() to guard against undersized buffer in
  protocol 2 path
- Add missing #include <stddef.h> in log.h and util.h in util.c
- Fix indentation in ctap_credential_management.h struct members

+++ fido2/inc/ctap_credential_management.h
@@ -49,13 +49,13 @@
 // clang-format on
 
 typedef struct {
-	int subCommand;
+	uint8_t subCommand;
 	struct {
 		uint8_t rpIdHash[32];
-        bool rpIdHash_present;
+		bool rpIdHash_present;
 		CTAP_credentialDescriptor credentialDescriptor;
-        bool credentialId_present;
-        bool credentialUser_present;
+		bool credentialId_present;
+		bool credentialUser_present;
 	} subCommandParams;
 
 	struct {
@@ -68,10 +68,10 @@ typedef struct {
 
 	uint8_t pinUvAuthParam[PIN_UV_AUTH_PARAM_MAX_SIZE];
 	uint8_t pinUvAuthParam_present;
-	int pinProtocol;
+	uint8_t pinProtocol;
 } CTAP_credMgmt;
 
 CtapStatus ctap_credential_management(CborEncoder *encoder, uint8_t *request,
-				      int length);
+				      size_t length);
 
 #endif
